test(breadcrumb): recategorize and clean up browser tests
